inserts patch pixel data into textures

+++ include/mapComponentStructs.h
@@ -39,7 +39,7 @@ typedef struct texture {
     char name[8];
     int16_t width;
     int16_t height;
-    int8_t* pixels;
+    uint32_t* pixels;
 } texture;
 
 #endif //DOOMLOOKER_MAPCOMPONENTSTRUCTS_H
